fix(executor): increase gemini-cli legacy initial stall timeout to 600s
Fixes #1619 where gemini-cli sessions were killed prematurely with
'gemini_legacy_initial_stall: no stdout within 600s'.

Root cause: classify_gemini_legacy_initial_stall() hardcoded a 120s
default timeout, while CSA pipeline defines 600s as the default for
gemini-cli. This inconsistency caused misleading error messages.

Changes:
- Add DEFAULT_GEMINI_LEGACY_INITIAL_RESPONSE_TIMEOUT_SECONDS = 600
- Update classify_gemini_legacy_initial_stall to use the new constant
- Add test to verify 600s default timeout behavior
- Refactor: extract OAuth functions to transport_gemini_oauth.rs module
  to meet 800-line file size limit (675 lines now)
- Bump version to 0.1.803

The new timeout aligns with pipeline.rs DEFAULT_GEMINI_INITIAL_RESPONSE_TIMEOUT_SECONDS
and provides sufficient time for gemini-cli OAuth auth and model loading.
