+++ crates/csa-executor/src/lib.rs
@@ -13,6 +13,7 @@ pub mod model_spec;
 pub mod session_config;
 pub mod session_id;
 pub mod transport;
+pub(crate) mod transport_gemini_oauth;
 pub(crate) mod transport_gemini_retry;
 pub mod transport_openai_compat;
 pub mod transport_tmux;
