+++ crates/csa-executor/src/transport_gemini_helpers.rs
@@ -12,13 +12,12 @@ use crate::executor::Executor;
 #[cfg(feature = "acp")]
 use crate::transport_gemini_retry::{gemini_retry_model, gemini_should_use_api_key};
 
-pub(crate) const GEMINI_OAUTH_PROMPT_SUMMARY: &str =
-    "gemini-cli auth failure: OAuth browser prompt detected; no tool output produced";
 pub(crate) const GEMINI_MCP_ISSUES_DETECTED_SUMMARY: &str =
     "MCP issues detected. Run /mcp list for status.";
 #[cfg(feature = "acp")]
 pub(crate) const GEMINI_ACP_INITIAL_STALL_REASON: &str = "gemini_acp_initial_stall";
 pub(crate) const GEMINI_LEGACY_INITIAL_STALL_REASON: &str = "gemini_legacy_initial_stall";
+const DEFAULT_GEMINI_LEGACY_INITIAL_RESPONSE_TIMEOUT_SECONDS: u64 = 600;
 #[cfg(feature = "acp")]
 const DEFAULT_GEMINI_ACP_INITIAL_RESPONSE_TIMEOUT_SECONDS: u64 = 180;
 #[cfg(feature = "acp")]
@@ -203,7 +202,8 @@ pub(super) fn classify_gemini_legacy_initial_stall(
 
     Some(GeminiLegacyInitialStallClassification {
         code: GEMINI_LEGACY_INITIAL_STALL_REASON,
-        timeout_seconds: timeout_seconds.unwrap_or(120),
+        timeout_seconds: timeout_seconds
+            .unwrap_or(DEFAULT_GEMINI_LEGACY_INITIAL_RESPONSE_TIMEOUT_SECONDS),
     })
 }
 
@@ -504,130 +504,6 @@ pub(super) fn apply_gemini_sandbox_runtime_contract(
     let env_overrides = gemini_sandbox_runtime_env_overrides(env);
     apply_gemini_sandbox_runtime_env_overrides(isolation_plan, &env_overrides);
     Ok(())
-}
-
-pub(crate) fn is_gemini_oauth_prompt_result(execution: &ExecutionResult) -> bool {
-    let stdout_has_auth_text =
-        execution.output.contains("authentication") || execution.output.contains("Authentication");
-    let stderr_has_auth_text = execution.stderr_output.contains("authentication")
-        || execution.stderr_output.contains("Authentication");
-    if !stdout_has_auth_text && !stderr_has_auth_text {
-        return false;
-    }
-
-    let normalized_stdout = if stdout_has_auth_text {
-        normalize_gemini_prompt_text(&execution.output)
-    } else {
-        String::new()
-    };
-    let normalized_stderr = if stderr_has_auth_text {
-        normalize_gemini_prompt_text(&execution.stderr_output)
-    } else {
-        String::new()
-    };
-    let combined = if normalized_stderr.is_empty() {
-        normalized_stdout.clone()
-    } else if normalized_stdout.is_empty() {
-        normalized_stderr.clone()
-    } else {
-        format!("{normalized_stdout}\n{normalized_stderr}")
-    };
-
-    if !contains_gemini_oauth_prompt(&combined) {
-        return false;
-    }
-
-    !combined.lines().any(|line| {
-        line.contains("\"type\":\"turn.completed\"")
-            || line.contains("\"type\": \"turn.completed\"")
-            || line.trim() == "turn.completed"
-    })
-}
-
-pub(crate) fn classify_gemini_oauth_prompt_result(execution: &mut ExecutionResult) {
-    execution.exit_code = 1;
-    execution.summary = GEMINI_OAUTH_PROMPT_SUMMARY.to_string();
-    if execution.stderr_output.is_empty() {
-        execution.stderr_output = GEMINI_OAUTH_PROMPT_SUMMARY.to_string();
-    } else if !execution
-        .stderr_output
-        .contains(GEMINI_OAUTH_PROMPT_SUMMARY)
-    {
-        if !execution.stderr_output.ends_with('\n') {
-            execution.stderr_output.push('\n');
-        }
-        execution
-            .stderr_output
-            .push_str(GEMINI_OAUTH_PROMPT_SUMMARY);
-    }
-}
-
-pub fn contains_gemini_oauth_prompt(text: &str) -> bool {
-    let lower = text.to_ascii_lowercase();
-    lower.contains("opening authentication page in your browser")
-        || (lower.contains("opening authentication page")
-            && lower.contains("do you want to continue"))
-        || (lower.contains("authentication page in your browser")
-            && lower.contains("do you want to continue"))
-}
-
-pub fn normalize_gemini_prompt_text(text: &str) -> String {
-    let mut cleaned = String::new();
-    let mut in_guard = false;
-    for raw_line in strip_ansi_escape_sequences(text).lines() {
-        let line = raw_line.trim_end_matches('\r');
-        let trimmed = line.trim();
-        if trimmed.starts_with("<csa-caller-sa-guard") {
-            in_guard = true;
-            continue;
-        }
-        if trimmed.starts_with("</csa-caller-sa-guard>") {
-            in_guard = false;
-            continue;
-        }
-        if trimmed.starts_with("<csa-caller-prompt-injection") {
-            in_guard = true;
-            continue;
-        }
-        if trimmed.starts_with("</csa-caller-prompt-injection>") {
-            in_guard = false;
-            continue;
-        }
-        if in_guard
-            || trimmed.is_empty()
-            || trimmed.starts_with("[csa-hook]")
-            || trimmed.starts_with("WARNING: weave.lock")
-            || trimmed.starts_with("csa run context:")
-            || trimmed.starts_with("Running scope as unit:")
-        {
-            continue;
-        }
-        let stripped = trimmed.strip_prefix("[stdout] ").unwrap_or(trimmed);
-        cleaned.push_str(stripped);
-        cleaned.push('\n');
-    }
-    cleaned
-}
-
-pub fn strip_ansi_escape_sequences(text: &str) -> String {
-    let mut stripped = String::with_capacity(text.len());
-    let mut chars = text.chars().peekable();
-    while let Some(ch) = chars.next() {
-        if ch != '\u{1b}' {
-            stripped.push(ch);
-            continue;
-        }
-        if !matches!(chars.peek(), Some('[')) {
-            continue;
-        }
-        let _ = chars.next();
-        for next in chars.by_ref() {
-            if ('@'..='~').contains(&next) {
-                break;
-            }
-        }
-    }
-    stripped
 }
 
 /// Convert a `tokio::task::JoinError` into a descriptive `anyhow::Error`.
